src/functions/requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-core